1.铺货前获取shopify的Collections、Type和Tags 2.铺货数据带上Collections、Type和Tags进行铺货 3.Shopify的订单下单的发货订单没有图片显示处理

+++ mall-portal/src/main/java/com/macro/mall/portal/util/OrderUtils.java
@@ -303,7 +303,7 @@ public class OrderUtils {
                 orderItemList.add(orderItem);
 
                 XmsCustomerSkuStock tempSkuStock = new XmsCustomerSkuStock();
-                tempSkuStock.setId(pmsSkuStock.getId().intValue());
+                tempSkuStock.setId(pmsSkuStock.getId());
                 tempSkuStock.setStock(pmsSkuStock.getStock());
                 skuStockInsertList.add(tempSkuStock);
             }
